:recycle: Rename data folder to data parsing scripts
rename data folder to data parsing scripts
